Refactor to create Officials objects in seperate files
