+++ src/Objects/Official.ts
@@ -1,0 +1,10 @@
+export interface Official {
+    role: string;
+    name: string;
+    country?: string;
+}
+
+export interface MatchOfficials {
+    matchId: string;
+    officials: Official[];
+}
